Add support for reading secrets one at a time
This explores adding an optional new `variables` parameter that will
read secrets out one by one into an ENV var that is visible in the
pipeline.yml.

The syntax and behaviour is modelled after the aws-ssm plugin. It can
be helpful in some cases for the ENV vars names to be listed explicitly
in the pipeline.yml - it makes their existence more obvious to
future maintainers, and it allows the value to be changed to read from
different secret keys without mutating the secret value.

This required making both styles optional, so I've taken the
controversial approach of removing the default `env` key from the
existing "batch" loading syntax. This would mean customers upgrading
from v0.1.0 to this might need to add `key: env` to keep the same
behaviour. I'm open to alternatives if we don't want a breaking change.